remove dflash v1 worker
